requirements.txt
Flask==2.3.3
pymysql==1.1.0
python-dotenv==1.0.0
boto3==1.28.85
botocore==1.31.85
APScheduler==3.10.4
bcrypt==4.0.1
pandas==2.1.1
openpyxl==3.1.2